requirements.txt
torch
torchvision
tqdm
prettytable
torchinfo==1.8.0
plyfile==0.9
trimesh
tensorboard
pyvista
rich==13.7.1
CythonGSL
loguru
torchdata
transformers
nltk
ipykernel
ipywidgets
pillow
accelerate
python-dotenv
gradio
open3d
bpy
openai
hydra-core 
datasets
matplotlib
seaborn
trl
peft
wandb==0.17.5
bitsandbytes
jsonargparse